requirements.txt
python-dotenv==1.0.0
python-binance==1.0.19
requests==2.31.0

